+++ packages/framework/src/Component/Elasticsearch/Debug/Exception/NotSupportedException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Component\Elasticsearch\Debug\Exception;
+
+use Exception;
+
+class NotSupportedException extends Exception
+{
+}