+++ packages/framework/src/Twig/ModuleExtension.php
@@ -11,7 +11,7 @@ class ModuleExtension extends Twig_Extension
     /**
      * @var \Shopsys\FrameworkBundle\Model\Module\ModuleFacade
      */
-    private $moduleFacade;
+    protected $moduleFacade;
 
     /**
      * @param \Shopsys\FrameworkBundle\Model\Module\ModuleFacade $moduleFacade