+++ project-base/src/Shopsys/ShopBundle/Form/Front/Customer/Password/NewPasswordFormType.php
@@ -22,7 +22,7 @@ class NewPasswordFormType extends AbstractType
             ->add('newPassword', RepeatedType::class, [
                 'type' => PasswordType::class,
                 'options' => [
-                    'attr' => ['autocomplete' => 'off'],
+                    'attr' => ['autocomplete' => 'new-password'],
                 ],
                 'first_options' => [
                     'constraints' => [